doc: move trott to tsc emeritus
(Effective after the next TSC meeting, November 21 2018.)

I'm going to step away from TSC stuff for about 6 weeks. No exciting
back-story here other than that I have a lot going on and I need a
break. And I've also said that people should do exactly this more often.
Maybe I'll make this an annual thing for me or something.

While the decision of if and when to restore my TSC status is entirely
up to the active members of the TSC, I intend to request to re-join in
January.

PR-URL: https://github.com/nodejs/node/pull/24492
Reviewed-By: Richard Lau <riclau@uk.ibm.com>
Reviewed-By: Gireesh Punathil <gpunathi@in.ibm.com>
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Michaël Zasso <targos@protonmail.com>
Reviewed-By: Anatoli Papirovski <apapirovski@mac.com>
Reviewed-By: Colin Ihrig <cjihrig@gmail.com>
Reviewed-By: Anna Henningsen <anna@addaleax.net>
Reviewed-By: Michael Dawson <michael_dawson@ca.ibm.com>
Reviewed-By: Joyee Cheung <joyeec9h3@gmail.com>
Reviewed-By: Franziska Hinkelmann <franziska.hinkelmann@gmail.com>
Reviewed-By: Weijia Wang <starkwang@126.com>
Reviewed-By: Tiancheng "Timothy" Gu <timothygu99@gmail.com>
Reviewed-By: Tobias Nießen <tniessen@tnie.de>